requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
httpx==0.25.2
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
langchain==0.1.0
langchain-core==0.1.0
langchain-anthropic==0.1.0
langgraph==0.1.0
mcp==0.3.0
typing-extensions==4.8.0
pytest==7.4.3
pytest-asyncio==0.21.1 